Change Mapping to 7.3 format Twitter Stream sentiment is now supported
