+++ src/startup.sh
@@ -4,8 +4,8 @@
 #stocksight is released under the Apache 2.0 license. See
 #LICENSE for the full license text.
 
-echo "Waiting for other dependent instances to spawn... (60 seconds)"
-sleep 60;
+echo "Waiting for other dependent instances to spawn... (30 seconds)"
+sleep 30;
 
 echo "Copy kibana dashboard if they don't exist";
 python import.kibana.py
@@ -14,13 +14,13 @@ tick_time=900
 tick=0
 let sentiment_time=900*4
 
-#echo "Spawning Tweet Sentiment receiver instance";
-#python tweet.sentiment.py &
+echo "Spawning Tweet Sentiment receiver instance";
+python tweet.sentiment.py &
 
 while true
 do
     echo "Spawning stock price receiver instance";
-    python stockprice.py &
+    python stockprice.docker.py &
     echo "Will get stock data again in ${tick_time} sec...";
     let tick_mod=tick%4
