+++ src/stockprice.docker.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-"""stockprice.py - get stock price from Yahoo and add to
+"""stockprice.docker.py - get stock price from Yahoo and add to
 Elasticsearch.
 See README.md or https://github.com/shirosaidev/stocksight
 for more information.
@@ -10,23 +10,29 @@ Copyright (C) Allen (Jian Feng) Xie 2019
 stocksight is released under the Apache 2.0 license. See
 LICENSE for the full license text.
 """
-
 import argparse
+import logging
 import sys
 
-from StockSight.NewsHeadlineListener import *
-from StockSight.EsMap.Sentiment import *
+from config import symbols
+from StockSight.Initializer.ElasticSearch import es
+from StockSight.EsMap.StockPrice import mapping
+from StockSight.StockPriceListener import StockPriceListener
 
 
-STOCKSIGHT_VERSION = '0.1-b.6'
+STOCKSIGHT_VERSION = '0.1-b.5'
 __version__ = STOCKSIGHT_VERSION
 
 
+
+
 if __name__ == '__main__':
+
     # parse cli args
     parser = argparse.ArgumentParser()
+
     parser.add_argument("-v", "--verbose", action="store_true",
-                       help="Increase output verbosity")
+                        help="Increase output verbosity")
     parser.add_argument("--debug", action="store_true",
                         help="Debug message output")
     parser.add_argument("-q", "--quiet", action="store_true",
@@ -36,6 +42,28 @@ if __name__ == '__main__':
                         help="Prints version and exits")
     args = parser.parse_args()
 
+    # set up logging
+    logger = logging.getLogger('stocksight')
+    logger.setLevel(logging.INFO)
+    eslogger = logging.getLogger('elasticsearch')
+    eslogger.setLevel(logging.WARNING)
+    requestslogger = logging.getLogger('requests')
+    requestslogger.setLevel(logging.WARNING)
+    logging.addLevelName(
+        logging.INFO, "\033[1;32m%s\033[1;0m"
+                      % logging.getLevelName(logging.INFO))
+    logging.addLevelName(
+        logging.WARNING, "\033[1;31m%s\033[1;0m"
+                         % logging.getLevelName(logging.WARNING))
+    logging.addLevelName(
+        logging.ERROR, "\033[1;41m%s\033[1;0m"
+                       % logging.getLevelName(logging.ERROR))
+    logging.addLevelName(
+        logging.DEBUG, "\033[1;33m%s\033[1;0m"
+                       % logging.getLevelName(logging.DEBUG))
+    logformatter = '%(asctime)s [%(levelname)s][%(name)s] %(message)s'
+    loglevel = logging.INFO
+    logging.basicConfig(format=logformatter, level=loglevel)
     if args.verbose:
         logger.setLevel(logging.INFO)
         eslogger.setLevel(logging.INFO)
@@ -53,19 +81,18 @@ if __name__ == '__main__':
         for symbol in symbols:
             try:
                 logger.info('Creating new Elasticsearch index or using existing ' + symbol)
-                es.indices.create(index="stocksight_"+symbol+"_sentiment", body=mapping, ignore=[400, 404])
-                url = "https://finance.yahoo.com/quote/%s/?p=%s" % (symbol, symbol)
+                es.indices.create(index="stocksight_"+symbol+"_price", body=mapping, ignore=[400, 404])
 
-                logger.info('NLTK tokens required: ' + str(nltk_tokens_required))
-                logger.info('NLTK tokens ignored: ' + str(nltk_tokens_ignored))
-                logger.info("Scraping news for %s from %s ..." % (symbol, url))
+                stockprice = StockPriceListener()
 
-                # create instance of NewsHeadlineListener
-                newslistener = NewsHeadlineListener(symbol, url)
+                stockprice.get_price(symbol=symbol)
             except Exception as e:
                 logger.warning("%s" % e)
                 pass
+        # get stock price
 
+    except Exception as e:
+        logger.warning("Exception: Failed to get stock data caused by: %s" % e)
     except KeyboardInterrupt:
         print("Ctrl-c keyboard interrupt, exiting...")
         sys.exit(0)
